Changed certificate validity from 1 year to 10 years
fast-ca is made for quick certificates, not for high-grade production servers, so the probability that the interest in a quick certificate server is high enough to crack a good certificate in 10 years should be slim. If new algorythms become standard or old get deprecated, one should create new certificates either way.